requirements.txt
flask>=2.3.0
torch>=1.9.0
torchaudio>=0.9.0
sounddevice>=0.4.4
numpy>=1.21.0
google-generativeai>=0.3.0
transformers>=4.21.0
gtts>=2.3.0
PyPDF2>=3.0.0
datasets>=2.0.0
scikit-learn>=1.1.0
nltk>=3.7
werkzeug>=2.3.0
pandas
numpy
